+++ tests/test_integration.py
@@ -1,12 +1,19 @@
+"""
+Test behaviors specific to importlib_metadata.
+
+These tests are excluded downstream in CPython as they
+test functionality only in importlib_metadata or require
+behaviors ('packaging') that aren't available in the
+stdlib.
+"""
+
 import unittest
 import packaging.requirements
 import packaging.version
 
 from . import fixtures
 from importlib_metadata import (
-    MetadataPathFinder,
     _compat,
-    distributions,
     version,
 )
 
@@ -29,10 +36,13 @@ class IntegrationTests(fixtures.DistInfoPkg, unittest.TestCase):
 
 class FinderTests(fixtures.Fixtures, unittest.TestCase):
     def test_finder_without_module(self):
-        class ModuleFreeFinder(fixtures.NullFinder):
+        class ModuleFreeFinder:
             """
             A finder without an __module__ attribute
             """
+
+            def find_module(self, name):
+                pass
 
             def __getattribute__(self, name):
                 if name == '__module__':
@@ -41,27 +51,3 @@ class FinderTests(fixtures.Fixtures, unittest.TestCase):
 
         self.fixtures.enter_context(fixtures.install_finder(ModuleFreeFinder()))
         _compat.disable_stdlib_finder()
-
-
-class DistSearch(unittest.TestCase):
-    def test_search_dist_dirs(self):
-        """
-        Pip needs the _search_paths interface to locate
-        distribution metadata dirs. Protect it for PyPA
-        use-cases (only). Ref python/importlib_metadata#111.
-        """
-        res = MetadataPathFinder._search_paths('any-name', [])
-        assert list(res) == []
-
-    def test_interleaved_discovery(self):
-        """
-        When the search is cached, it is
-        possible for searches to be interleaved, so make sure
-        those use-cases are safe.
-
-        Ref #293
-        """
-        dists = distributions()
-        next(dists)
-        version('importlib_metadata')
-        next(dists)